Fixup for #7, #9
This is my fixup for issues #7 and #9.

The problem was: When resizing the table, there is a new table allocated
on the stack of the function ht_resize(), which is then passed to another
function ht_insert_he(). This function takes the members of the
hash_table type and uses them for inserting the values. As they need to
be rehashed, it wants to use the hashfunctions as well. But these were
not passed to the new hashtable object!

This patch fixes this issue.