Fix up ht_destroy() and ht_resize()
This is a patch for consistancy: When destroying a hashtable, all of its
members are set to either zero or NULL. This patch adds the NULLifying
of the hashfunction pointers.

As the ht_resize function makes use of the ht_destroy() function, it has
to reset the hashfunction pointers to the hashtable object which got
resized.